upl/mouser.ccp, venture/spcforce.cpp, venture/suprridr.cpp: consolidated drivers in single files, minor cleanups
